requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.4.0
safetensors>=0.3.0

# Memory Optimization (for 24GB VRAM efficiency)
xformers>=0.0.20
bitsandbytes>=0.41.0  # Optional 8-bit optimizers

# Data & Processing
datasets>=2.14.0
Pillow>=9.5.0
opencv-python>=4.8.0
numpy>=1.24.0

# Training & Tracking
wandb>=0.15.0
tensorboard>=2.13.0
tqdm>=4.65.0

# Configuration
PyYAML>=6.0
pydantic>=2.0.0

# Development
jupyterlab>=4.0.0
pytest>=7.4.0
black>=23.0.0
python-dotenv>=1.0.0